Sprint 6 - Button ad insta-story is always visible for user. User can close application in onbackPressed() if he is on home Fragment.
