+++ .idea/misc.xml
@@ -14,6 +14,7 @@
         <entry key="..\:/Android/InstaWSBFinal/app/src/main/res/layout/image_items.xml" value="0.20625" />
         <entry key="..\:/Android/InstaWSBFinal/app/src/main/res/layout/notification_items.xml" value="0.36666666666666664" />
         <entry key="..\:/Android/InstaWSBFinal/app/src/main/res/layout/profile_image_items.xml" value="0.36666666666666664" />
+        <entry key="..\:/Android/InstaWSBFinal/app/src/main/res/layout/stories_layout.xml" value="0.36666666666666664" />
         <entry key="..\:/Android/InstaWSBFinal/app/src/main/res/layout/user_items.xml" value="0.20625" />
         <entry key="..\:/Android/ZPwIT_WSB_GR1_project/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.3828125" />
         <entry key="..\:/Android/ZPwIT_WSB_GR1_project/app/src/main/res/drawable/ic_launcher_background.xml" value="0.3828125" />
