+++ app/src/main/java/com/zpwit_wsb_gr1_project/adapter/StoriesAdapter.java
@@ -52,62 +52,22 @@ public class StoriesAdapter extends RecyclerView.Adapter<StoriesAdapter.StoriesH
     @Override
     public void onBindViewHolder(@NonNull StoriesAdapter.StoriesHolder holder, @SuppressLint("RecyclerView") int position) {
 
-        if (position == 0) {
+        Glide.with(activity)
+                .load(list.get(position).getUrl())
+                .timeout(6500)
+                .into(holder.imageView);
 
-            Glide.with(activity)
-                    .load(activity.getResources().getDrawable(R.drawable.ic_add))
-                    .into(holder.imageView);
+        holder.imageView.setOnClickListener(v -> {
 
-            holder.imageView.setOnClickListener(v ->
-                    activity.startActivity(new Intent(activity, StoryAddActivity.class)));
+            //open story
+            Intent intent = new Intent(activity, ViewStoryActivity.class);
+            intent.putExtra(VIDEO_URL_KEY, list.get(position).getUrl());
+            intent.putExtra(FILE_TYPE, list.get(position).getType());
+            activity.startActivity(intent);
 
-        }else {
 
-            Glide.with(activity)
-                    .load(list.get(position).getUrl())
-                    .timeout(6500)
-                    .into(holder.imageView);
 
-            holder.imageView.setOnClickListener(v -> {
-
-                if (holder.getAbsoluteAdapterPosition() == 0) {
-                    //new story
-
-                    Dexter.withContext(activity)
-                            .withPermissions(Manifest.permission.READ_EXTERNAL_STORAGE,
-                                    Manifest.permission.WRITE_EXTERNAL_STORAGE)
-                            .withListener(new MultiplePermissionsListener() {
-                                @Override
-                                public void onPermissionsChecked(MultiplePermissionsReport multiplePermissionsReport) {
-
-                                    if (multiplePermissionsReport.areAllPermissionsGranted()) {
-
-                                        activity.startActivity(new Intent(activity, StoryAddActivity.class));
-
-                                    } else {
-                                        Toast.makeText(activity, "Please allow permission from settings.", Toast.LENGTH_SHORT).show();
-                                    }
-
-                                }
-
-                                @Override
-                                public void onPermissionRationaleShouldBeShown(List<PermissionRequest> list, PermissionToken permissionToken) {
-                                    permissionToken.continuePermissionRequest();
-                                }
-                            }).check();
-
-                } else {
-                    //open story
-                    Intent intent = new Intent(activity, ViewStoryActivity.class);
-                    intent.putExtra(VIDEO_URL_KEY, list.get(position).getUrl());
-                    intent.putExtra(FILE_TYPE, list.get(position).getType());
-                    activity.startActivity(intent);
-
-                }
-
-            });
-
-        }
+        });
 
     }
