+++ app/src/main/res/layout/stories_layout.xml
@@ -9,6 +9,7 @@
         android:layout_height="@dimen/_38mdp"
         app:civ_border_color="@android:color/darker_gray"
         app:civ_border_width="2dp"
+        android:layout_margin="@dimen/_1mdp"
         android:id="@+id/imageView" />
